+++ RELEASENOTES.md
@@ -5,6 +5,7 @@ _Salesforce Lightning Design System :: React Components :: design-system-react_
 
 These are changes that have backwards-compatible solutions present and that compatibiity will be removed at a breaking change release in the future.
 
+- `Checkbox`'s `onChange` now recieves `event, {checked: [boolean], indeterminate: [boolean] }`. Previously, `checked` was the first parameter.
 - `TabsPane` has be replaced with `TabsPanel`.
 - `Input`'s props: `iconPosition`, `iconAssistiveText`, `iconCategory`, `iconName`, `onIconClick` are deprecated. An `Icon` component should be used instead.
 - `DataTable`'s' `collection`, `onSelect`, `onDeselect` are deprecated.